Rename internal logging module to fix import issues with standard logging module
